price with BigDecimal, parking handler name changed, change to LocalDateTime and formatter, static block initalizer redundent, use block instead.

+++ src/parkinglot/strategy/FourWheelerWeekday.java
@@ -2,10 +2,12 @@ package parkinglot.strategy;
 
 import parkinglot.model.Ticket;
 
+import java.math.BigDecimal;
+
 public class FourWheelerWeekday implements CostStrategy{
 
     @Override
-    public double calculateCost(Ticket ticket) {
+    public BigDecimal calculateCost(Ticket ticket) {
         return getDefaultPrice(ticket);
     }
 }
